replacing PhantomJS with headless FF
